+++ tests/index.js
@@ -9,6 +9,7 @@ import * as overlappingMarks from './overlapping-marks.test.js'
 import * as customCompare from './custom-compare.test.js'
 import * as inlineNodes from './inline-nodes.test.js'
 import * as attrAttribution from './attr-attribution.test.js'
+import * as brokenSchema from './broken-schema.test.js'
 import * as prosemirrorRdt from './prosemirror-rdt.test.js'
 import * as swallowFormats from './swallow-formats.test.js'
 // import * as tr from './tr.test.js'
@@ -31,6 +32,7 @@ runTests({
   customCompare,
   inlineNodes,
   attrAttribution,
+  brokenSchema,
   prosemirrorRdt,
   swallowFormats
   // prosemirror,