+++ tests/index.node.js
@@ -16,6 +16,7 @@ import * as overlappingMarks from './overlapping-marks.test.js'
 import * as customCompare from './custom-compare.test.js'
 import * as inlineNodes from './inline-nodes.test.js'
 import * as attrAttribution from './attr-attribution.test.js'
+import * as brokenSchema from './broken-schema.test.js'
 import * as ySyncRdt from './y-sync-rdt.test.js'
 import * as prosemirrorRdt from './prosemirror-rdt.test.js'
 import * as swallowFormats from './swallow-formats.test.js'
@@ -73,6 +74,7 @@ runTests({
   customCompare,
   inlineNodes,
   attrAttribution,
+  brokenSchema,
   ySyncRdt,
   prosemirrorRdt,
   swallowFormats